LogFile

+++ Solo-feladat.BLL/Dtos/Coordinate.cs
@@ -9,7 +9,8 @@ namespace Solo_feladat.BLL.Dtos
         public Guid Id { get; set; }
         public Guid FlightId { get; set; }
         public Guid AirportId { get; set; }
-        public float LatitudeCoord { get; set; }
-        public float LongitudeCoord { get; set; }
+
+        public double LatitudeCoord { get; set; }
+        public double LongitudeCoord { get; set; }
     }
 }
